# sql/distsqlrun: refactor mergeJoiner to implement RowSource