The performance diff at smaller input sizes is due to the removal of the
`log.VEventf` lines.

```
name                           old time/op    new time/op    delta
MergeJoiner/InputSize=0-8        3.39µs ± 2%    2.30µs ± 2%  -32.22%  (p=0.000 n=10+9)
MergeJoiner/InputSize=4-8        5.66µs ± 1%    4.19µs ± 1%  -25.91%  (p=0.000 n=10+8)
MergeJoiner/InputSize=16-8       8.76µs ± 1%    7.14µs ± 1%  -18.55%  (p=0.000 n=9+10)
MergeJoiner/InputSize=256-8      66.6µs ± 1%    64.3µs ± 1%   -3.50%  (p=0.000 n=9+9)
MergeJoiner/InputSize=4096-8      987µs ± 1%     991µs ± 1%     ~     (p=0.089 n=10+10)
MergeJoiner/InputSize=65536-8    16.1ms ± 3%    16.4ms ± 3%     ~     (p=0.113 n=9+10)
```

Release note: None