@petermattis 2018-01-04
I broke this change into a series of smaller commits so that the reviewer can see the transformations I made before arriving at the final state.