@petermattis 2018-01-04
> The performance diff at smaller input sizes is due to the removal of the log.VEventf lines.

This actually isn't the cause of the performance delta. It is something to do with the last commit. I'm investigating.