@petermattis 2018-01-04
The difference is actually due to not using `DrainAndClose`. That method spawns a goroutine if there is more than one input to drain:

```
		for _, input := range srcs[1:] {
			wg.Add(1)
			go func(input RowSource) {
				DrainAndForwardMetadata(ctx, input, dst)
				wg.Done()
			}(input)
		}
```

The refactored `mergeJoiner` drains inputs serially. I'm mildly surprised this makes a difference, but there you go.