@petermattis 2018-01-05
> This should probably be a FastIntSet (though I don't think that should be part of this PR, just pointing it out)

Heh, I noticed that too. Do you have any sense for how big typical merge-join batches are?

@arjunravinarayan you need to acknowledge the changes you requested.

---

Review status: 0 of 6 files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21219#-:-L26y4n2V2QNuAQxNaot:b-twpqtg)*
<!-- Sent from Reviewable.io -->