@RaduBerinde 2018-01-05
*[pkg/sql/distsqlrun/mergejoiner.go, line 167 at r7](https://reviewable.io:443/reviews/cockroachdb/cockroach/21219#-L1yhb4Z_L3bGrHloPE0-r7-167:-L271N5vgd_Apr5GBzyL:b-p61ik9) ([raw file](https://github.com/cockroachdb/cockroach/blob/914565af6dbe75112928aafd87fb18d12c214f19/pkg/sql/distsqlrun/mergejoiner.go#L167)):*
<details><summary><i>Previously, arjunravinarayan (Arjun Narayan) wrote…</i></summary><blockquote>

Elegant! Thanks for the comment. Would be good to get one more review on the refactored merge joiner state machine. @RaduBerinde, do you mind taking a quick look at this function?
</blockquote></details>

It looks good to me. This part is easier to review with the separate commits.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21219)*
<!-- Sent from Reviewable.io -->