@RaduBerinde 2018-01-05
It's hard to give a blanket statement, but it's common for the merge-join batches to be small (e.g. 1-row batches when joining on a key).

---







*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21219#-:-L271aOi4YMub-KAZPsc:b5bt9cz)*
<!-- Sent from Reviewable.io -->